tests/wpt/tests/scroll-animations/animation-trigger/parsing/timeline-trigger-exit-range-start-invalid.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-animations-2/#animation-trigger">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("timeline-trigger-exit-range-start", "peek 50%");
test_invalid_value("timeline-trigger-exit-range-start", "50% contain");
test_invalid_value("timeline-trigger-exit-range-start", "50% cover");
test_invalid_value("timeline-trigger-exit-range-start", "50% entry");
test_invalid_value("timeline-trigger-exit-range-start", "50% enter");
test_invalid_value("timeline-trigger-exit-range-start", "50% exit");
test_invalid_value("timeline-trigger-exit-range-start", "contain contain");
test_invalid_value("timeline-trigger-exit-range-start", "none");
test_invalid_value("timeline-trigger-exit-range-start", "cover 50% enter 50%");
test_invalid_value("timeline-trigger-exit-range-start", "auto contain");
</script>
